Fetch pagination feature with refactoring
