+++ sailor/_base/__init__.py
@@ -1,1 +1,2 @@
 from .masterdata import MasterDataField, MasterDataEntity, MasterDataEntitySet, add_properties   # noqa: F401
+from .fetch import fetch_data, parse_filter_parameters, apply_filters_post_request  # noqa: F401
